Updated the Login and Register for both Students and Instructor, with Firestore db
